pyproject.toml
[tool.poetry]
name = "spend-analyzer"
version = "0.1.0"
description = ""
authors = ["Colton Voege <colton@beacons.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mypy = "^1.8.0"
google-api-python-client = "^2.116.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
inflection = "^0.5.1"
ruff = "^0.2.0"
black = "^24.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
